Rework Dn to be usable to find ip addresses
Ready for finding ip Addresses

+++ client/impl/src/main/java/it/xeniaprogetti/cisco/ucs/plugin/client/impl/model/org/root/LsServer.java
@@ -29,7 +29,7 @@ public class LsServer extends UcsElement {
     @JacksonXmlProperty(isAttribute = true)
     public String extIPState;
     @JacksonXmlProperty(isAttribute = true)
-    public int fltAggr;
+    public long fltAggr;
     @JacksonXmlProperty(isAttribute = true)
     public String fsmDescr;
     @JacksonXmlProperty(isAttribute = true)
